# ui/client/shared/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateProjectRequest": {
        "description": "Request body for creating a project.",
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "contract": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Contract"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "kind": {
            "$ref": "#/components/schemas/ProjectKind",
            "default": "side_project"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "weekly_goal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weekly Goal"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateProjectRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProjectKind": {
        "description": "What a project is to the person tracking it. A contract is read only on a day job.",
        "enum": [
          "day_job",
          "freelance",
          "side_project"
        ],
        "title": "ProjectKind",
        "type": "string"
      },
...
      "RecordTimeRequest": {
        "description": "Request body for recording time (start/stop timer).",
        "properties": {
          "time": {
            "format": "date-time",
            "title": "Time",
            "type": "string"
          }
        },
        "title": "RecordTimeRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "A time tracking application",
    "title": "Beats API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/projects/": {
      "get": {
        "description": "List projects, optionally augmented with aggregation fields.\n\nBackward compatible: without `include`, the response carries the same\nfields as ProjectResponse plus null aggregation slots \u2014 older clients\nthat ignore unknown null fields keep working.\n\nP3.0 of the project-management revamp. First cut loops per-project\nserver-side; if profiling shows it's needed, a follow-up (P3.0b)\nmoves to a Mongo $facet pipeline for true batched aggregation.",
        "operationId": "list_projects_api_projects__get",
        "parameters": [
          {
            "in": "query",
            "name": "archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Archived",
              "type": "boolean"
            }
          },
          {
            "description": "Comma-separated aggregations to populate per project. Supported: 'totals' (total_minutes), 'this_week' (weekly_minutes + effective_goal trio, and on a day job the contract's expected / worked / remaining / balance), 'last_tracked' (last_tracked_at). Omitted \u21d2 slim response.",
            "in": "query",
            "name": "include",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated aggregations to populate per project. Supported: 'totals' (total_minutes), 'this_week' (weekly_minutes + effective_goal trio, and on a day job the contract's expected / worked / remaining / balance), 'last_tracked' (last_tracked_at). Omitted \u21d2 slim response.",
              "title": "Include"
            }
          },
          {
            "description": "IANA timezone name (e.g. America/New_York). Defaults to UTC.",
            "in": "query",
            "name": "tz",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "IANA timezone name (e.g. America/New_York). Defaults to UTC.",
              "title": "Tz"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectsListItemResponse"
                  },
                  "title": "Response List Projects Api Projects  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Projects",
        "tags": [
          "Projects"
        ]
      },
      "post": {
        "description": "Create a new project.",
        "operationId": "create_project_api_projects__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Project",
        "tags": [
          "Projects"
        ]
      },
      "put": {
        "description": "Update an existing project.",
        "operationId": "update_project_api_projects__put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Project",
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/projects/stop": {
      "post": {
        "description": "Stop the currently running timer.",
        "operationId": "end_project_timer_api_projects_stop_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordTimeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "End Project Timer",
        "tags": [
          "Projects"
        ]
      }
    },
...
    "/api/projects/{project_id}/archive": {
      "post": {
        "description": "Archive a project.",
        "operationId": "archive_project_api_projects__project_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Project",
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/projects/{project_id}/contract": {
      "put": {
        "description": "Replace a day job's whole contract: terms, holiday region, opening\nbalance and end date, in one shape, like goal-overrides. Leaves `kind`\nalone; on a project that is not a day job it is a 409 (NOT_A_DAY_JOB),\nand a region the calendar does not know is a 400 (UNKNOWN_HOLIDAY_REGION).",
        "operationId": "replace_contract_api_projects__project_id__contract_put",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Contract"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replace Contract",
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/projects/{project_id}/contract/week": {
      "get": {
        "description": "One week of a day job against its contract \u2014 expected, worked,\nremaining, the running balance as of today, and the seven days with\ntheir holiday and absence. Worked hours are bucketed by the local day\neach beat started on, in `tz`; a running timer counts up to now.\n409 on a project that is not a day job (NOT_A_DAY_JOB) or has no\ncontract yet (NO_CONTRACT).",
        "operationId": "get_contract_week_api_projects__project_id__contract_week_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "description": "Monday of the week wanted. Defaults to the current week in `tz`.",
            "in": "query",
            "name": "week_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Monday of the week wanted. Defaults to the current week in `tz`.",
              "title": "Week Of"
            }
          },
          {
            "description": "IANA timezone name (e.g. America/New_York). Defaults to UTC.",
            "in": "query",
            "name": "tz",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "IANA timezone name (e.g. America/New_York). Defaults to UTC.",
              "title": "Tz"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractWeek"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Contract Week",
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/projects/{project_id}/daily-average": {
      "get": {
        "description": "Get the average daily session time for a project over the last 30 days.",
        "operationId": "get_daily_average_api_projects__project_id__daily_average_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Daily Average",
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/projects/{project_id}/git-activity": {
      "get": {
        "description": "Get GitHub commit counts correlated with a project's tracked time.",
        "operationId": "get_git_activity_api_projects__project_id__git_activity_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "format": "date",
              "title": "Start",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "format": "date",
              "title": "End",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Git Activity",
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/projects/{project_id}/goal-overrides": {
      "put": {
        "description": "Replace goal overrides for a project.",
        "operationId": "update_goal_overrides_api_projects__project_id__goal_overrides_put",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/GoalOverrideRequest"
                },
                "title": "Overrides",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Goal Overrides",
        "tags": [
          "Projects"
        ]
      }
    },
...
    "/api/projects/{project_id}/start": {
      "post": {
        "description": "Start a timer for a project.",
        "operationId": "start_project_timer_api_projects__project_id__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordTimeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Project Timer",
        "tags": [
          "Projects"
        ]
      }
    },
...
  }
}